--- a/src/abbrev.rs
+++ b/src/abbrev.rs
@@ -95,10 +95,7 @@ impl Abbreviations {
         let mut abbrevs = Abbreviations::empty();
 
         loop {
-            let result = Abbreviation::parse_null(input).map(|(rest, _)| (rest, None));
-            let result =
-                result.or_else(|_| Abbreviation::parse(input).map(|(rest, a)| (rest, Some(a))));
-            let (rest, abbrev) = try!(result);
+            let (rest, abbrev) = try!(Abbreviation::parse(input));
             input = rest;
 
             match abbrev {
@@ -169,16 +166,6 @@ impl Abbreviation {
         &self.attributes[..]
     }
 
-    /// Parse an abbreviation's code.
-    fn parse_code(input: &[u8]) -> ParseResult<(&[u8], u64)> {
-        let (rest, code) = try!(parse_unsigned_leb(input));
-        if code == 0 {
-            Err(Error::AbbreviationCodeZero)
-        } else {
-            Ok((rest, code))
-        }
-    }
-
     /// Parse an abbreviation's tag.
     fn parse_tag(input: &[u8]) -> ParseResult<(&[u8], constants::DwTag)> {
         let (rest, val) = try!(parse_unsigned_leb(input));
@@ -223,24 +210,19 @@ impl Abbreviation {
         Ok((input, attrs))
     }
 
-    /// Parse a non-null abbreviation.
-    fn parse(input: &[u8]) -> ParseResult<(&[u8], Abbreviation)> {
-        let (rest, code) = try!(Self::parse_code(input));
+    /// Parse an abbreviation. Return `None` for the null abbreviation, `Some`
+    /// for an actual abbreviation.
+    fn parse(input: &[u8]) -> ParseResult<(&[u8], Option<Abbreviation>)> {
+        let (rest, code) = try!(parse_unsigned_leb(input));
+        if code == 0 {
+            return Ok((rest, None));
+        }
+
         let (rest, tag) = try!(Self::parse_tag(rest));
         let (rest, has_children) = try!(Self::parse_has_children(rest));
         let (rest, attributes) = try!(Self::parse_attributes(rest));
         let abbrev = Abbreviation::new(code, tag, has_children, attributes);
-        Ok((rest, abbrev))
-    }
-
-    /// Parse a null abbreviation.
-    fn parse_null(input: &[u8]) -> ParseResult<(&[u8], ())> {
-        let (rest, code) = try!(parse_unsigned_leb(input));
-        if code == 0 {
-            Ok((rest, ()))
-        } else {
-            Err(Error::ExpectedZero)
-        }
+        Ok((rest, Some(abbrev)))
     }
 }
 
@@ -573,23 +555,6 @@ mod tests {
     }
 
     #[test]
-    fn test_parse_abbreviation_code_ok() {
-        let buf = [0x01, 0x02];
-        let (rest, code) = Abbreviation::parse_code(&buf).expect("Should parse code");
-        assert_eq!(code, 0x01);
-        assert_eq!(rest, &buf[1..]);
-    }
-
-    #[test]
-    fn test_parse_abbreviation_code_zero() {
-        let buf = [0x00];
-        match Abbreviation::parse_code(&buf) {
-            Err(Error::AbbreviationCodeZero) => {}
-            otherwise => panic!("Unexpected result: {:?}", otherwise),
-        };
-    }
-
-    #[test]
     fn test_parse_abbreviation_tag_ok() {
         let buf = [0x01, 0x02];
         let (rest, tag) = Abbreviation::parse_tag(&buf).expect("Should parse tag");
@@ -645,11 +610,17 @@ mod tests {
             0x04
         ];
 
-        let expect = Abbreviation::new(
-            1, constants::DW_TAG_subprogram, constants::DW_CHILDREN_no,
-            vec![
-                AttributeSpecification::new(constants::DW_AT_name, constants::DW_FORM_string),
-            ]);
+        let expect = Some(
+            Abbreviation::new(
+                1,
+                constants::DW_TAG_subprogram,
+                constants::DW_CHILDREN_no,
+                vec![
+                    AttributeSpecification::new(constants::DW_AT_name,
+                                                constants::DW_FORM_string),
+                ]
+            )
+        );
 
         let (rest, abbrev) = Abbreviation::parse(&buf).expect("Should parse abbreviation");
         assert_eq!(abbrev, expect);
@@ -670,17 +641,9 @@ mod tests {
             0x04
         ];
 
-        let (rest, _) = Abbreviation::parse_null(&buf).expect("Should parse null abbreviation");
+        let (rest, abbrev) = Abbreviation::parse(&buf).expect("Should parse null abbreviation");
+        assert!(abbrev.is_none());
         assert_eq!(rest, [0x01, 0x02, 0x03, 0x04]);
-    }
-
-    #[test]
-    fn test_parse_null_abbreviation_nonzero() {
-        let buf = [0x01];
-        match Abbreviation::parse_null(&buf) {
-            Err(Error::ExpectedZero) => {}
-            otherwise => panic!("Unexpected result: {:?}", otherwise),
-        };
     }
 
     #[test]