Before:
```
test bench_parsing_debug_abbrev ... bench:      33,643 ns/iter (+/- 6,243)
test bench_parsing_debug_info   ... bench:   5,914,364 ns/iter (+/- 819,837)
```

After:
```
test bench_parsing_debug_abbrev ... bench:      29,905 ns/iter (+/- 7,143)
test bench_parsing_debug_info   ... bench:   5,913,803 ns/iter (+/- 894,585)
```

Looks like a little bit of a speed up for parsing abbreviations, but its within the margin of error. Either way, I think this actually cleans up the loop-until-we-parse-a-null-thing code a bit.

What do you think @philipc?