@philipc 2016-07-26
Nice cleanup