@fitzgen 2016-07-27
Thanks for checking this out! I removed the unused error variants, and I'm going to merge now.